% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_tryCatchLog_result.R
\name{append.to.last.tryCatchLog.result}
\alias{append.to.last.tryCatchLog.result}
\title{Appends a new log entry to the stored logging output of the last call to \code{tryCatchLog} or \code{tryLog}}
\usage{
append.to.last.tryCatchLog.result(new.log.entry)
}
\arguments{
\item{new.log.entry}{the new log entry (normally as \code{character} string)}
}
\value{
the extended logging result of the last call to \code{tryCatchLog} or \code{tryLog} as \code{list}
}
\description{
You can get the last logging output by calling \code{\link{last.tryCatchLog.result}}.
}
\details{
THIS FUNCTION IS USED ONLY PACKAGE INTERNALLY!
}
