% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLogMessage.R
\name{buildLogMessage}
\alias{buildLogMessage}
\title{Internal helper function to build a log message}
\usage{
buildLogMessage(log.message, call.stack, omit.last.items = 0)
}
\arguments{
\item{log.message}{a text message}

\item{call.stack}{a call stack created by \code{\link{sys.calls}}}

\item{omit.last.items}{number of stack trace items to ignore (= last x items)}
}
\value{
A ready to use log message with a pretty printed compact and detailled stack trace
              (as \code{character})
}
\description{
Combines a log message with a compact and a detailled stack trace with the option
              to ignore the last x stack trace items (normally created due to internal error handling
              and therefore irrelevant for the user).
}
\note{
THIS IS A PACKAGE INTERNAL FUNCTION AND THEREFORE NOT EXPORTED.
}
