% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_add_row.R
\name{config.add.row}
\alias{config.add.row}
\title{Add a new entry (row) to an existing configuration}
\usage{
config.add.row(
  config,
  cond.class,
  silent = FALSE,
  write.to.log = TRUE,
  log.as.severity = tryCatchLog::Severity.Levels$INFO,
  include.full.call.stack = FALSE,
  include.compact.call.stack = TRUE
)
}
\arguments{
\item{config}{an existing configuration}

\item{cond.class}{the class name of the condition as character (eg. "error")}

\item{silent}{\code{\link{logical}}: \code{TRUE} = do not propagate the condition to other registered handlers (= "muffle" in R speak).}

\item{write.to.log}{\code{\link{logical}}: \code{TRUE} = Call the logging function for the caught condition}

\item{log.as.severity}{Severity level for the \code{cond.class} (use the constants from \code{\link{Severity.Levels}} or the equivalent character strings)}

\item{include.full.call.stack}{\code{\link{logical}}: Shall the full call stack be included in the log output?}

\item{include.compact.call.stack}{\code{\link{logical}}: Shall the compact call stack (including only calls with source code references)
be included in the log output?}
}
\value{
The new config that includes the new row(s)
}
\description{
You could also add multiple rows in one call by passing vectors of the same length for each argument
(except for the \code{config} argument of course).
}
\details{
The config passed in as argument is NOT modified, you have to reassign the return value for this (see examples).

For a detailed description of the argument semantics see the related function \code{\link{config.create}}.
}
\examples{
config <- config.create()
config <- config.add.row(config, "myInfo")
config <- config.add.row(config, "myError", FALSE, TRUE, Severity.Levels$FATAL, TRUE, TRUE)
options("tryCatchLog.global.config" = config)
}
