% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_load_save.R
\name{config.load}
\alias{config.load}
\title{Load a configuration to control the behavior of \code{tryCatchLogExt()} from a config file}
\usage{
config.load(file.name = "tryCatchLog_config.csv")
}
\arguments{
\item{file.name}{The name of a file that contains a valid configuration}
}
\value{
A configuration
}
\description{
Load a configuration to control the behavior of \code{tryCatchLogExt()} from a config file
}
\examples{
config.save(config.create())
config.load()
}
\seealso{
\code{\link{config.create}}, \code{\link{config.save}}
}
