% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_load_save.R
\name{config.save}
\alias{config.save}
\title{Save a configuration to control the behavior of \code{tryCatchLogExt()} into a config file}
\usage{
config.save(config, file.name = "tryCatchLog_config.csv")
}
\arguments{
\item{config}{A configuration (created via \code{\link{config.create}}}

\item{file.name}{A file name for the configuration file}
}
\value{
\code{file.name} (invisible)
}
\description{
Save a configuration to control the behavior of \code{tryCatchLogExt()} into a config file
}
\examples{
config.save(config.create())
}
\seealso{
\code{\link{config.create}}, \code{\link{config.load}}
}
