% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_newline.R
\name{determine.platform.NewLine}
\alias{determine.platform.NewLine}
\title{Determines the operating system specific new line character(s)}
\usage{
determine.platform.NewLine()
}
\value{
the new line character(s) for the current operating system
}
\description{
CR + LF on Windows, else only LF...
}
\details{
This function is pendant to Microsoft's .Net "Environment.NewLine".
}
\note{
THIS IS A PACKAGE INTERNAL FUNCTION AND THEREFORE NOT EXPORTED.
}
\references{
\url{https://stackoverflow.com/questions/47478498/build-string-with-os-specific-newline-characters-crlf-lf-cr-to-write-it-into}
}
