% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pretty_tryCatchLog_options.R
\name{get.pretty.option.value}
\alias{get.pretty.option.value}
\title{gets the current value of an option as key/value string}
\usage{
get.pretty.option.value(option.name)
}
\arguments{
\item{option.name}{Name of the option (as character)}
}
\value{
The option as key/value string in one line
}
\description{
The data type is also indicated if an option is set (since a wrong data type may cause problems).
If an option is not set "(not set)" is shown as value.
}
\details{
THIS IS AN INTERNAL PRIVATE FUNCTION OF THE PACKAGE.
}
\examples{
\dontrun{
tryCatchLog:::get.pretty.option.value("warn")
# [1] "Option warn = 0 (double)" }
}
\seealso{
\code{\link{get.pretty.tryCatchLog.options}}
}
