% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pretty_tryCatchLog_options.R
\name{get.pretty.tryCatchLog.options}
\alias{get.pretty.tryCatchLog.options}
\title{Gets the current option values of all options supported by the `tryCatchLog` package}
\usage{
get.pretty.tryCatchLog.options()
}
\value{
The current option settings as string (one per line as key/value pair), e. g.
\preformatted{
Option tryCatchLog.write.error.dump.file = FALSE (logical)
Option tryCatchLog.write.error.folder = . (character)
Option tryCatchLog.silent.warnings = FALSE (logical)
Option tryCatchLog.silent.messages = (not set)
}
}
\description{
This is a convenience function whose result can be used e. g. to log the current settings.
}
\details{
If an option is not set the string "(not set)" is shown as value.

The data type is also indicated if an option is set (since a wrong data type may cause problems).
}
\examples{
cat(get.pretty.tryCatchLog.options())  # "cat" does apply new line escape characters

}
