% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_already_logged.R
\name{is.already.logged}
\alias{is.already.logged}
\title{Check if a condition is already logged in the \code{\link{last.tryCatchLog.result}}}
\usage{
is.already.logged(msg.text, full.stack.trace)
}
\arguments{
\item{msg.text}{the condition message as it would be logged}

\item{full.stack.trace}{the call stack of the condition}
}
\value{
\code{TRUE} if the condition is already contained in the current log, else \code{FALSE}
}
\description{
Check if a condition is already logged in the \code{\link{last.tryCatchLog.result}}
}
\note{
Introduced as replacement for \code{is.duplicated.log.entry} due to
      a required signature change to fix issue #55
      (https://github.com/aryoda/tryCatchLog/issues/55).
}
\seealso{
\code{\link{last.tryCatchLog.result}},
         \code{\link{build.log.entry}}
}
