% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package_available.R
\name{is.package.available}
\alias{is.package.available}
\title{Checks if a package is installed and can be loaded}
\usage{
is.package.available(package.name)
}
\arguments{
\item{package.name}{Name of the package (as string)}
}
\value{
\code{TRUE} if the packages is installed, otherwise \code{FALSE} (invisible)

http://r-pkgs.had.co.nz/description.html
}
\description{
Use this function to check for optional package dependencies
within this package.
}
\details{
This is a package-internal function!

See section ‘Good practice’ in '?.onAttach'.
}
\examples{
tryCatchLog:::is.package.available("tryCatchLog")  # must be TRUE :-)

}
