% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_newline.R
\name{is.windows}
\alias{is.windows}
\title{Determines if R is running on a Windows operating system}
\usage{
is.windows()
}
\value{
\code{TRUE} of running on a Windows OS else \code{FALSE}
}
\description{
Throws a warning if an indication for Windows OS were found but the Windows OS cannot be recognized for sure
(via a second different check).
}
\examples{
is.windows()
}
