% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_tryCatchLog_log.R
\name{last.tryCatchLog.log}
\alias{last.tryCatchLog.log}
\title{Gets the logging result of the last call to \code{tryCatchLog} or \code{tryLog}}
\usage{
last.tryCatchLog.log()
}
\value{
the logging result of the last call to \code{tryCatchLog} or \code{tryLog} as \code{list}
        comprised of one element per logged condition
}
\description{
The last logging result after calling \code{tryCatchLog} or `\code{tryLog} can be retrieved by
calling this function.
}
\details{
The typical use case is to get and store the log output not only in a log file but
also in another place that is not supported by \code{futile.logger}, e. g. in
a data base table of your application or displaying it in a GUI (user interface).

Another use case is to review the last log output on the console during debugging.
}
\examples{
    \\code{last.tryCatchLog.log()}

}
