% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2console.R
\name{log2console}
\alias{log2console}
\title{Prints a time-stamped log message to the console incl. the severity level}
\usage{
log2console(severity.level, msg)
}
\arguments{
\item{severity.level}{String containing the severity level
(\code{ERROR}, \code{WARN} or \code{INFO}) of the message}

\item{msg}{The message to be printed (as character).}
}
\value{
The log message as it was printed to the console.
                       \code{NA} is printed as empty string.
}
\description{
This is a package-internal function.
}
\examples{
tryCatchLog:::log2console("WARN", "this is my last warning")

}
