% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_newline.R
\name{platform.NewLine}
\alias{platform.NewLine}
\title{Gets the operating system specific new line character(s)}
\usage{
platform.NewLine()
}
\value{
the new line character(s) for the current operating system
}
\description{
CR + LF on Windows, else only LF...
}
\details{
The newline character(s) are determined once at package loading time.

You normally don't need to care about this since "\\n" in strings works
on every platform (operating system) because it is converted automatically
when printed, saved in or loaded from a file.

A typical use case is a unit test that compares an actual file
(eg. with "\\r\\n" on Windows) to an expected file that was eg. created on Linux
with "\\n" platform newline and you do byte compare of the files
(which will fail unless to "undo" the platform newline differences).
}
\examples{
platform.NewLine()
}
