% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_tryCatchLog_log.R
\name{reset.last.tryCatchLog.log}
\alias{reset.last.tryCatchLog.log}
\title{Resets the stored logging output of the last call to \code{tryCatchLog} or \code{tryLog} to an empty list}
\usage{
reset.last.tryCatchLog.log()
}
\value{
invisible: TRUE
}
\description{
You can get the last logging output by calling \code{\link{last.tryCatchLog.log}}.
}
\details{
THIS FUNCTION IS USED ONLY PACKAGE INTERNALLY!
}
\examples{
    \\code{reset.last.tryCatchLog.log(NULL)}
}
\seealso{
\code{\link{last.tryCatchLog.log}}
}
