% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_tryCatchLog_result.R
\name{reset.last.tryCatchLog.result}
\alias{reset.last.tryCatchLog.result}
\title{Resets the stored logging output of the last call to \code{tryCatchLog} or \code{tryLog} to an empty list}
\usage{
reset.last.tryCatchLog.result()
}
\value{
invisible: TRUE
}
\description{
You can get the last logging output by calling \code{\link{last.tryCatchLog.result}}.
}
\note{
THIS IS A PACKAGE INTERNAL FUNCTION AND THEREFORE NOT EXPORTED.
}
\seealso{
\code{\link{last.tryCatchLog.result}},
         \code{\link{append.to.last.tryCatchLog.result}},
}
