% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_log_entry.R
\name{build.log.entry}
\alias{build.log.entry}
\title{Creates a log entry row containing all relevant logging information in columns}
\usage{
build.log.entry(severity, log.message, call.stack, omit.last.items = 0)
}
\arguments{
\item{severity}{severity level of the log entry ((ERROR, WARNING, MESSAGE etc.)}

\item{call.stack}{a call stack created by \code{\link{sys.calls}}}

\item{omit.last.items}{the number of stack trace items to ignore (= last x calls) in
the passed \code{call.stack} since they are caused by using \code{tryCatchLog}}

\item{msg.text}{the message text of the thrown condition}
}
\value{
A \code{\link{data.frame}} with one and the following columns:
        \enumerate{
        \item{severity    - the serverity level of the log entry (ERROR, WARNING, MESSAGE)}
        \item{log.message - the message text of the log entry}
        \item{compact.stack.trace - the short stack trace containing only entries with source code
                                    references down to line of code that has thrown the condition}
        \item{full.stack.trace    - the full stack trace with all calls down to the line of code that
                                    has thrown the condition (including calls to R internal functions
                                    and other functions even when the source code in not available).}
        }
}
\description{
The serverity level should correspond to the condition class.
}
\note{
THIS IS A PACKAGE INTERNAL FUNCTION AND THEREFORE NOT EXPORTED.
}
