% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_log_output.R
\name{build.log.output}
\alias{build.log.output}
\title{Internal helper function to build a decent log output for logging}
\usage{
build.log.output(log.entry)
}
\arguments{
\item{log.entry}{A \code{data.frame} with one row (created by \code{\link{build.log.entry}}
containing the logging information to be prepared for the logging output}
}
\value{
A ready to use log message with a pretty printed compact and detailled stack trace
              (as \code{character})
}
\description{
Combines a log message with a compact and a detailled stack trace
}
\note{
THIS IS A PACKAGE INTERNAL FUNCTION AND THEREFORE NOT EXPORTED.
}
