% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_newline.R
\name{platform.NewLine}
\alias{platform.NewLine}
\title{Gets the operating system specific new line character(s)}
\usage{
platform.NewLine()
}
\value{
the new line character(s) for the current operating system
}
\description{
CR + LF on Windows, else only LF...
}
\details{
The newline character(s) are determined once at package loading time.
}
\examples{
platform.NewLine()
}
